Android Wear notifications
One stack group element per contact is displayed.
Messages appear in the order they were received (latest message bottom).

Notifications are ordered by the time of latest message arrival.
Contacts with the newest message appear on top.

Bug on Android 5.0 and 5.1

If an alarm (ringtone, vibration) is set on the summary notification,
every single thread message appears on the handheld.
If an alarm is set on the single thread notifications, the more
notifications arrive, the more alarms are kicked off.

To work around the issue, the summary notification's priority is set to
'default' instead of 'high', which suppresses the heads-up notification.
This notification is shown in the Android notification bar nevertheless.

Users enabling this workaround will often read notifications on the
wearable and therefore not miss the heads-up notification.

On older / newer devices, this setting has no effect, the summary
notification is displayed as usual.

See https://code.google.com/p/android/issues/detail?id=130689

Developed with Sebastian Chlan <sebastian.chlan@gmail.com>

// FREEBIE